Last commit for working feature

+++ app/src/main/java/com/alwan/plantdisease/util/Extension.kt
@@ -67,4 +67,4 @@ fun Uri.toFile(context: Context): File {
     return myFile
 }
 
-fun Double?.orZero() = this ?: 0.0
+fun Float?.orZero() = this ?: 0f
